Move VesselReportingList/Current to parent in Frontend
