+++ frontend/src/features/Reporting/components/Current/utils.ts
